requirements.txt
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
blessed==1.19.1
boto3==1.21.18
botocore==1.24.18
cached-property==1.5.2
cachetools==5.0.0
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.3
cryptography==36.0.1
defusedxml==0.7.1
Django==2.2
django-admin-honeypot==1.1.0
django-admin-rangefilter==0.8.4
django-admin-thumbnails==0.2.6
django-allauth==0.48.0
django-appconf==1.0.5
django-cleanup==6.0.0
django-nested-inline==0.4.4
django-storages==1.12.3
docker==3.7.3
docker-compose==1.25.5
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
google-api-core==2.5.0
google-api-python-client==2.38.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.0
googleapis-common-protos==1.55.0
httplib2==0.20.4
idna==2.7
importlib-metadata==4.11.2
jmespath==0.10.0
jsonschema==2.6.0
oauthlib==3.2.0
paramiko==2.9.2
pathspec==0.9.0
Pillow==9.0.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.3.1
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
s3transfer==0.5.2
semantic-version==2.8.5
six==1.12.0
sqlparse==0.4.2
termcolor==1.1.0
texttable==0.9.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.1.9
websocket-client==0.59.0
zipp==3.7.0
